Allow .NET Exceptions to be passed to JavaScript
Currently if JavaScript code calls into a .NET method which produces
an exception, the exception is not passed to the calling code in
JavaScript, instead it is thrown out of the RunModule() call.

This commit allows .NET Exceptions to be wrapped with JavaScript Error
objects containing the exception message. In case the Error is not
handled in JavaScript code, the original .NET exception is extracted
from the Error and is rethrown in the .NET code.